requirements.txt
numpy
matplotlib
scipy
pyvisa
Pillow
